+++ repos/ktor/client/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/client/key_value/KtorStandartKeyValueRepo.kt
@@ -1,27 +1,40 @@
 package dev.inmo.micro_utils.repos.ktor.client.key_value
 
+import dev.inmo.micro_utils.ktor.client.UnifiedRequester
+import dev.inmo.micro_utils.ktor.common.StandardKtorSerialFormat
+import dev.inmo.micro_utils.ktor.common.standardKtorSerialFormat
 import dev.inmo.micro_utils.repos.*
 import io.ktor.client.HttpClient
-import kotlinx.serialization.KSerializer
+import kotlinx.serialization.*
 
 class KtorStandartKeyValueRepo<K, V> (
     baseUrl: String,
     baseSubpart: String,
-    client: HttpClient = HttpClient(),
+    unifiedRequester: UnifiedRequester,
     keySerializer: KSerializer<K>,
     valueSerializer: KSerializer<V>,
     valueNullableSerializer: KSerializer<V?>
 ) : StandardKeyValueRepo<K, V>,
     ReadStandardKeyValueRepo<K, V> by KtorReadStandardKeyValueRepo(
         "$baseUrl/$baseSubpart",
-        client,
+        unifiedRequester,
         keySerializer,
         valueSerializer,
         valueNullableSerializer
     ),
     WriteStandardKeyValueRepo<K, V> by KtorWriteStandardKeyValueRepo(
         "$baseUrl/$baseSubpart",
-        client,
+        unifiedRequester,
         keySerializer,
         valueSerializer
-    )
+    ) {
+    constructor(
+        baseUrl: String,
+        baseSubpart: String,
+        client: HttpClient = HttpClient(),
+        keySerializer: KSerializer<K>,
+        valueSerializer: KSerializer<V>,
+        valueNullableSerializer: KSerializer<V?>,
+        serialFormat: StandardKtorSerialFormat = standardKtorSerialFormat
+    ) : this(baseUrl, baseSubpart, UnifiedRequester(client, serialFormat), keySerializer, valueSerializer, valueNullableSerializer)
+}